+++ chroniclepy/run.py
@@ -62,6 +62,9 @@ def get_parser():
         help = 'What time does daytime start?  In 24h format (eg. 10:00)')
     summaryargs.add_argument('--nighttime', action='store', default= '22:00',
         help = 'What time does nighttime start?  In 24h format (eg. 22:00)')
+    summaryargs.add_argument('--maxdays', action='store', default=None, type=int,
+        help = "What is the maximum number of days you want to analyze per person?"
+    )
     return parser
 
 def main():
@@ -109,7 +112,8 @@ def main():
             weekdefinition = opts.weekdefinition,
             splitday = opts.splitday,
             daytime = opts.daytime,
-            nighttime = opts.nighttime
+            nighttime = opts.nighttime,
+            maxdays = opts.maxdays
         )
 
 if __name__ == '__main__':